# api-spec/api.yaml
openapi: 3.1.0
info:
  title: Nearby Mechanic Finder API
  version: 1.0.0
  description: API for connecting vehicle owners with nearby mechanics.
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /mechanics:
    get:
      tags:
        - Mechanics
      summary: Find mechanics with optional location-based filtering
      parameters:
        - in: query
          name: latitude
          schema:
            type: number
            minimum: -90
            maximum: 90
          required: false
          description: Latitude for location-based search (default 0)
        - in: query
          name: longitude
          schema:
            type: number
            minimum: -180
            maximum: 180
          required: false
          description: Longitude for location-based search (default 0)
        - in: query
          name: radius
          schema:
            type: number
            minimum: 0
            maximum: 20000
          required: false
          description: Search radius in kilometers (default 10)
        - in: query
          name: serviceType
          schema:
            type: string
            enum:
              - "Tire Change"
              - "Oil Change"
              - "Battery Replacement"
              - "Engine Repair"
              - "Brake Service"
              - "Towing"
              - "Jump Start"
              - "Fuel Delivery"
              - "Lockout Service"
          required: false
          description: Filter mechanics by service type
        - in: query
          name: useLocation
          schema:
            type: boolean
          required: false
          description: Whether to use location-based filtering (default false)
      responses:
        "200":
          description: List of mechanics retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MechanicProfile"
  /mechanics/profile:
    get:
      tags:
        - Mechanics
      summary: Get current mechanic profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Mechanic profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MechanicProfile"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden - Not a mechanic
        "404":
          description: Profile not found
    post:
      tags:
        - Mechanics
      summary: Create or update mechanic profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MechanicProfileRequest"
      responses:
        "200":
          description: Mechanic profile created or updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MechanicProfile"
        "400":
          description: Invalid input
        "401":
          description: Unauthorized
        "403":
          description: Forbidden - Not a mechanic
  /mechanics/location:
    put:
      tags:
        - Mechanics
      summary: Update mechanic's current location and availability
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMechanicLocationRequest"
      responses:
        "200":
          description: Location updated successfully
        "401":
          description: Unauthorized
        "403":
          description: Forbidden - Not a mechanic
        "404":
          description: Mechanic profile not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MechanicProfile:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        services:
          type: array
          items:
            type: string
        availability:
          type: boolean
        rating:
          type: number
        certifications:
          type: array
          items:
            type: string
        experienceYears:
          type: integer
        currentLocation:
          type: object
          properties:
            latitude:
              type: number
            longitude:
              type: number
        lastLocationUpdate:
          type: string
          format: date-time
        serviceRadiusKm:
          type: number
          default: 10.0
    MechanicProfileRequest:
      type: object
      properties:
        services:
          type: array
          items:
            type: string
            enum:
              [
                "Tire Change",
                "Oil Change",
                "Battery Replacement",
                "Engine Repair",
                "Brake Service",
                "Towing",
                "Jump Start",
                "Fuel Delivery",
                "Lockout Service",
              ]
          minItems: 1
        availability:
          type: boolean
        certifications:
          type: array
          items:
            type: string
        experienceYears:
          type: integer
          minimum: 0
          maximum: 50
        serviceRadiusKm:
          type: number
          minimum: 1
          maximum: 100
          default: 10
      required:
        - services
        - availability
        - experienceYears
    UpdateMechanicLocationRequest:
      type: object
      properties:
        latitude:
          type: number
          minimum: -90
          maximum: 90
        longitude:
          type: number
          minimum: -180
          maximum: 180
        availability:
          type: boolean
      required:
        - latitude
        - longitude
        - availability
...